Fix no code coverage for namespaces/packages
Files that belong to a particular PHP namespace are represented in the
clover.xml code coverage file as being in a "package", and these files
couldn't be found with the xpath query being used previously. This
change should find all <file> elements in the file, allowing code
coverage data to be displayed for these files.